+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Known issues
 - notebook search with `m` or `y` in key won't work, as vscode configured default shortcut `m` to change cell to markdown and is annoying. [remove](https://code.visualstudio.com/docs/getstarted/keybindings#_keyboard-shortcuts-editor) `m` and `y` shortcuts for clean experience.
 
+
+## 1.0.45:
+- Create menu in explorer easy http file and notebook file
+- Update `dotextensions-build` to version 0.0.44a11
+  - error will fail at the end of resolution of all properties
+
+
 ## 1.0.44:
 - Update `dotextensions-build` to version 0.0.44a10
 - Enhance content resolution by including resolved property